feat(gitlab): Implement GitLab OAuth token refresh and improve integration
This change introduces automatic OAuth token refreshing for GitLab to handle expiring access tokens, ensuring a more stable authenticated session. It also includes several fixes and improvements to make GitLab integration more robust and feature-complete.

**Key Changes:**

**Authentication:**
- Implemented an `OAuthTokenRefresher` to automatically refresh GitLab access tokens before they expire.
- The `TokenDataSource` now includes a `refreshIfNeeded` method, which is called before making authenticated API requests to GitLab.
- Added `GITLAB_CLIENT_SECRET` to the build configuration to support the token refresh flow.
- The authentication dialog now provides clearer messages for expired tokens, prompting the user to sign in again.

**GitLab API Integration:**
- Refactored `DetailsRepository` to use platform-specific (GitHub/GitLab) implementations for fetching repository details, releases, READMEs, and user profiles.
- Correctly implemented API calls to fetch repository details by ID, latest releases, README files, and user profiles from the GitLab API.
- Improved the mapping from `GitLabProjectNetworkModel` to the internal `GithubRepoSummary` model, ensuring consistent data handling across platforms.
- Fixed an issue where the wrong token data source was used for authenticated GitLab HTTP clients.
- Navigation now passes the current API platform to the details screen to ensure the correct repository data is fetched.

**Minor Fixes & Refinements:**
- The repository ID type has been changed from `Int` to `Long` to support GitLab's larger ID values.
- Updated the authentication screen to display the GitLab icon when logging into GitLab.
- Refactored Koin dependency injection modules for better organization, moving token-related dependencies into a new `tokenModule`.

+++ composeApp/src/commonMain/kotlin/zed/rainxch/githubstore/app/navigation/GithubStoreGraph.kt
@@ -21,7 +21,7 @@ sealed interface GithubStoreGraph : NavKey {
 
     @Serializable
     data class DetailsScreen(
-        val repositoryId: Int
+        val repositoryId: Long
     ) : GithubStoreGraph
 
     @Serializable
